+++ js/upload-success.js
@@ -1,0 +1,24 @@
+import {uploadOverlay, uploadWrapper, uploadImageInput, clearUpload} from './upload-image.js';
+const successTemplate = document.querySelector('#success')
+  .content
+  .querySelector('.success');
+
+const showSuccess = () => {
+  const successElement = successTemplate.cloneNode(true);
+  const successButton = successElement.querySelector('.success__button');
+  uploadWrapper.classList.add('hidden');
+
+  const removeSuccessMessage = () => {
+    successButton.removeEventListener('click', removeSuccessMessage);
+    uploadWrapper.classList.remove('hidden');
+    successElement.remove();
+    uploadImageInput.value = '';
+    clearUpload();
+  };
+
+  successButton.addEventListener('click', removeSuccessMessage);
+
+  uploadOverlay.appendChild(successElement);
+};
+
+export {showSuccess};
